Move execution payload to core
